pyproject.toml
[tool.poetry]
name = "freelance-api"
version = "0.1.0"
description = ""
authors = ["Artur Boyun <arturboyun@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
aiosqlite = "^0.19.0"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
pytest-asyncio = "^0.23.4"
celery = "^5.3.6"
flower = "^2.0.1"
redis = "^5.0.1"
python-multipart = "^0.0.9"
asyncpg = "^0.29.0"
argon2-cffi = "^23.1.0"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
edgedb = "^1.9.0"


[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^24.1.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
